Refactor: changed RTResult placement
RTResult was living in lunfardo_parser, which doesn't make any sense, so now lives in its own file at "Lunfardo/src/rtresult"
